shcov: Add handling of lines which exceed PS4 limit
Issue #6: The bash PS4 environment variables has a 100 character limit (at
least in the 3.X version I'm using, which is the only version I've tested.)
This causes various tracebacks to occur in the code if a directory path is
deeper than that, since the parts list ends up having too few elements.

This commit adds a check to ensure that the parts list has four element, and
simply skips the line (after spewing an error message) if it does not. This
will cause inaccurate coverage, but at least the script will not simply abort.